+++ types/organization.go
@@ -1,11 +1,14 @@
 package types
 
-import "errors"
+import (
+	fmt "fmt"
+	"net/url"
+)
 
 // Validate returns an error if the organization does not pass validation tests
 func (o *Organization) Validate() error {
 	if err := ValidateName(o.Name); err != nil {
-		return errors.New("organization name " + err.Error())
+		return fmt.Errorf("organization name %s", err)
 	}
 
 	return nil
@@ -16,4 +19,9 @@ func FixtureOrganization(name string) *Organization {
 	return &Organization{
 		Name: name,
 	}
+}
+
+// URIPath returns the path component of a Organization URI.
+func (o *Organization) URIPath() string {
+	return fmt.Sprintf("/rbac/organizations/%s", url.PathEscape(o.Name))
 }